Separate command by chat types
