+++ modules/bot/controllers/privates/DefaultController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\bot\controllers;
+namespace app\modules\bot\controllers\privates;
 
 use app\modules\bot\components\Controller as Controller;
