+++ modules/bot/controllers/privates/My_genderController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\bot\controllers;
+namespace app\modules\bot\controllers\privates;
 
 use TelegramBot\Api\Types\Inline\InlineKeyboardMarkup;
 use Yii;
