+++ modules/bot/controllers/privates/My_profileController.php
@@ -1,26 +1,30 @@
 <?php
 
-namespace app\modules\bot\controllers;
+namespace app\modules\bot\controllers\privates;
 
 use \app\modules\bot\components\response\SendMessageCommand;
 use app\modules\bot\components\Controller as Controller;
 
 /**
- * Class HelpController
+ * Class My_profileController
  *
- * @package app\controllers\bot
+ * @package app\modules\bot\controllers
  */
-class HelpController extends Controller
+class My_profileController extends Controller
 {
     /**
      * @return array
      */
     public function actionIndex()
     {
+        $update = $this->getUpdate();
+
         return [
             new SendMessageCommand(
                 $this->getTelegramChat()->chat_id,
-                $this->render('index'),
+                $this->render('index', [
+                    'profile' => $update->getMessage()->getFrom(),
+                ]),
                 [
                     'parseMode' => $this->textFormat,
                 ]
