+++ modules/bot/controllers/privates/Random_intController.php
@@ -1,25 +1,24 @@
 <?php
 
-namespace app\modules\bot\controllers;
+namespace app\modules\bot\controllers\privates;
 
 use Yii;
 use app\modules\bot\components\Controller as Controller;
 
 /**
- * Class ReverseController
+ * Class Random_intController
  *
  * @package app\modules\bot\controllers
  */
-class ReverseController extends Controller
+class Random_intController extends Controller
 {
     /**
      * @return string
      */
     public function actionIndex($message = '')
     {
-        //TODO add reverse for $$message
-        return $message ? $message : '';
-
+        //TODO add flexible int min and max from $message
+        return random_int(1, 10);
         //return $this->render('index');
     }
 }
