+++ modules/bot/controllers/privates/Random_stringController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\bot\controllers;
+namespace app\modules\bot\controllers\privates;
 
 use Yii;
 use app\modules\bot\components\Controller as Controller;
