+++ modules/bot/controllers/privates/ReverseController.php
@@ -1,24 +1,25 @@
 <?php
 
-namespace app\modules\bot\controllers;
+namespace app\modules\bot\controllers\privates;
 
 use Yii;
 use app\modules\bot\components\Controller as Controller;
 
 /**
- * Class Random_intController
+ * Class ReverseController
  *
  * @package app\modules\bot\controllers
  */
-class Random_intController extends Controller
+class ReverseController extends Controller
 {
     /**
      * @return string
      */
     public function actionIndex($message = '')
     {
-        //TODO add flexible int min and max from $message
-        return random_int(1, 10);
+        //TODO add reverse for $$message
+        return $message ? $message : '';
+
         //return $this->render('index');
     }
 }
