+++ modules/bot/views/privates/my_birthday/index.php
