+++ modules/bot/views/privates/my_currency/currency-list.php
