+++ modules/bot/views/privates/my_currency/index.php
