+++ modules/bot/views/privates/my_email/merge-accounts.php
