+++ modules/bot/views/privates/my_location/footer.php
