+++ modules/bot/views/privates/my_location/index.php
