+++ modules/bot/views/privates/my_location/send-location.php
