+++ modules/bot/views/privates/my_location/update.php
