+++ modules/bot/views/privates/my_rating/index.php
